refactor(widgets,shared): chandge handleClick function, change import from shared
