+++ src/shared/constants/index.ts
@@ -1,2 +1,3 @@
 export * from "./navigation";
 export * from "./contacts";
+export * from "./html-attrs";
